+++ ShennonFano/ShennonFano.cpp
@@ -272,7 +272,7 @@ int main(void) {
 
     // Частотный анализ
     map<char, int> freqTable;
-    for (int i = 0; i < input.size(); i++) {
+    for (size_t i = 0; i < input.size(); i++) {
         freqTable[input[i]] += 1;
     }
 
@@ -296,7 +296,7 @@ int main(void) {
 
     // Строим map {символ->код}
     map<char, string> encodingTable;
-    for (int i = 0; i < chars.size(); i++) {
+    for (size_t i = 0; i < chars.size(); i++) {
         vector<int> code;
         BSTNode::BSTConnector* currentEdge = tree->search(chars[i]);
 
@@ -319,7 +319,7 @@ int main(void) {
     string encode = input;
 
     int position = 0;
-    for (int i = 0; i < input.size(); i++) {
+    for (size_t i = 0; i < input.size(); i++) {
         encode.replace(position, 1, encodingTable[input[i]]);
         position += encodingTable[input[i]].size();
     }
@@ -334,7 +334,7 @@ int main(void) {
 
     string selection;
     int selectionStart = 0;
-    for (int i = 0; i < encode.size(); i++) {
+    for (size_t i = 0; i < encode.size(); i++) {
         selection += encode[i];
         map<string, char>::const_iterator search = decodingTable.find(selection);
